refactor(core/layout): port-placement into auto-placement; drop unused constants
(1) port-placement.ts and its test move from layout/ root into
the flat-tree algorithm directory:

  layout/port-placement.ts       -> layout/auto-placement/flat-tree/port-placement.ts
  layout/port-placement.test.ts  -> layout/auto-placement/flat-tree/port-placement.test.ts

The file's content is algorithm-internal (decidePortSides reads
direction + device-type heuristics; placePorts uses the
algorithm's port-side decisions). It belongs next to the
algorithm that owns it. Internal imports adjusted; layout/
index.ts re-export updated to point at the new path.

The public surface is unchanged — placePorts is still imported
from @shumoku/core directly.

(3) Unused constants removed from core/constants.ts:

  - SMALL_LABEL_CHAR_WIDTH (6.5 px/char estimate)
  - ESTIMATED_CHAR_WIDTH   (7 px/char estimate)

Both were replaced by engine.text.measure when renderers
migrated in #295. No remaining consumers in libs/ or apps/.

173 / 173 tests pass.
